backend/requirements.txt
Django==5.2.0
requests==2.32.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
pillow==10.3.0